feat(lightcurve): Removed redundant buttons, and simplified lightcurve component.
