+++ components/cardLightCurve.vue
@@ -1,42 +1,27 @@
+<script
+  src="https://unpkg.com/htmx.org@1.9.6"
+  integrity="sha384-FhXw7b6AlE/jyjlZH5iHa/tTe9EpJ1Y55RjcgPbjeWMskSxZt1v9qkxLJWNJaGni"
+  crossorigin="anonymous"
+></script>
 <template>
   <v-col :cols="cols" :lg="lg" :md="md" :sm="sm">
     <v-card :class="cardClass">
-      <v-card-text>
-        <plots-light-curve-plot-htmx
-          :type="selected"
-          :object-id="objectId"
-          :dark="isDark"
-        />
-      </v-card-text>
-      <v-card width="100%">
-        <!-- OPTIONS -->
-        <v-card-actions class="py-0">
-          <!--RADIO BUTTONS-->
-          <buttons-lightcurve-radio-buttons
-            v-model="selected"
-            :options="options"
-          />
-          <v-spacer />
-          <v-row>
-            <!-- LIGHTCURVE BUTTONS -->
-            <v-col class="py-1">
-              <buttons-display-data-release
-                v-model="dataReleaseValues"
-                :datarelease="dataRelease"
-                :loading="isLoadingDataRelease"
-                :plot="selected"
-                @update-plot="updatePlotSelected"
-              />
-            </v-col>
-            <v-col class="py-1">
-              <buttons-download-lightcurve-button
-                :oid="objectId"
-                :detections="lightcurve.detections"
-                :non-detections="lightcurve.nonDetections"
-              />
-            </v-col>
-          </v-row>
-        </v-card-actions>
+      <v-card>
+        <v-card id="lightcurve-container" width="100%"> </v-card>
+        <v-card v-if="isLoading || error">
+          <v-card-text v-if="isLoading">
+            <v-progress-circular
+              indeterminate
+              color="primary"
+            ></v-progress-circular>
+            Fetching data for object {{ objectId }} ...
+          </v-card-text>
+          <v-card-text v-if="error">
+            <v-alert text prominent type="error" icon="mdi-cloud-alert">
+              {{ error }}
+            </v-alert>
+          </v-card-text>
+        </v-card>
       </v-card>
     </v-card>
   </v-col>
@@ -44,6 +29,7 @@
 
 <script>
 import { Vue, Component, Prop, Watch } from 'nuxt-property-decorator'
+import * as htmx from 'htmx.org'
 
 @Component
 export default class CardLightCurve extends Vue {
@@ -55,58 +41,15 @@ export default class CardLightCurve extends Vue {
 
   @Prop({ type: Number | String, default: 12 }) sm
 
-  @Prop({ type: Number, default: null }) period
-
   @Prop({ type: Boolean, default: true }) show
 
   @Prop({ type: String }) cardClass
 
-  selected = ''
-
-  dataReleaseValues = []
-
-  options = [
-    {
-      text: 'Difference Magnitude',
-      value: 'difference',
-      tooltip:
-        'The difference Magnitude light curve, is the absolute difference between a science and reference magnitudes.',
-      show: true,
-      default: true,
-    },
-    {
-      text: 'Apparent Magnitude',
-      value: 'apparent',
-      tooltip:
-        'Apparent magnitude light curve results from adding/subtracting the fluxes from the reference and difference in the same unit system and then converting to magnitudes.',
-      show: true,
-    },
-    {
-      text: 'Folded',
-      value: 'folded',
-      tooltip:
-        'The Period folded light curve, where time is transformed to time modulo the period (Phase).',
-      show: true,
-    },
-  ]
+  isLoading = true
+  error = ''
 
   get isDark() {
     return this.$vuetify.theme.isDark
-  }
-
-  get lightcurve() {
-    return {
-      detections: this.$store.state.lightcurve.detections,
-      nonDetections: this.$store.state.lightcurve.nonDetections,
-    }
-  }
-
-  get dataRelease() {
-    return this.$store.state.datarelease.dataReleaseLightCurve
-  }
-
-  get isLoadingDataRelease() {
-    return this.$store.state.datarelease.loading
   }
 
   get objectInformation() {
@@ -117,54 +60,41 @@ export default class CardLightCurve extends Vue {
     return this.$store.state.object.objectId
   }
 
-  @Watch('objectInformation')
-  onObjectInformation(val) {
-    this.options.forEach((x) => {
-      switch (x.value) {
-        case 'difference':
-          x.default = !val.corrected
-          break
-        case 'apparent':
-          x.show = val.corrected
-          x.default = val.corrected
-          break
-        case 'folded':
-          x.show = this.period !== null && val.corrected
-          break
-      }
-    })
+  updatePlotSelected(event) {
+    this.selected = event
   }
 
-  @Watch('dataRelease')
-  onDataReleaseValues(val) {
-    this.options.forEach((x) => {
-      switch (x.value) {
-        case 'apparent':
-          x.show = this.objectInformation.corrected || val.length > 0
-          x.default = this.objectInformation.corrected
-          break
-        case 'folded':
-          x.show =
-            this.period !== null &&
-            (this.objectInformation.corrected || val.length > 0)
-          break
+  mounted() {
+    this.$el.addEventListener('htmx:responseError', (event) => {
+      this.error = event.detail.error
+      this.isLoading = false
+    })
+    this.$el.addEventListener('htmx:afterRequest', (event) => {
+      if (event.detail.successful) {
+        this.error = ''
+        this.isLoading = false
       }
     })
-  }
-
-  @Watch('period')
-  onPeriod(val) {
-    if (this.objectInformation) {
-      this.options[2].show = val !== null && this.objectInformation.corrected
-    }
   }
 
   onDetectionClick(val) {
     if (val) this.$store.dispatch('lightcurve/changeDetection', val.index)
   }
 
-  updatePlotSelected(event) {
-    this.selected = event
+  @Watch('objectId', { immediate: true })
+  onIdChange(newId) {
+    this.error = ''
+    this.isLoading = true
+    if (newId) {
+      const url = `${this.$config.lightCurveApiv2Url}/htmx/lightcurve?oid=${newId}`
+      console.log(url)
+      const myDiv = document.getElementById('lightcurve-container')
+      if (myDiv) {
+        myDiv.innerHTML = `<div hx-get=${url} hx-trigger="updateLightcurve from:body"></div>`
+        htmx.process(myDiv)
+        document.body.dispatchEvent(new Event('updateLightcurve'))
+      }
+    }
   }
 }
 </script>
